+++ src/classes/babies/DolefulBaby.ts
@@ -1,4 +1,5 @@
 import {
+  CollectibleType,
   ModCallback,
   TearFlag,
   TearVariant,
@@ -8,6 +9,11 @@ import { Baby } from "../Baby";
 
 /** Starts with Soy Milk + booger tears. */
 export class DolefulBaby extends Baby {
+  /** Certain collectibles do not work with the baby effect. */
+  override isValid(player: EntityPlayer): boolean {
+    return !player.HasCollectible(CollectibleType.TRISAGION);
+  }
+
   @Callback(ModCallback.POST_FIRE_TEAR)
   postFireTear(tear: EntityTear): void {
     tear.ChangeVariant(TearVariant.BOOGER);