+++ docs/history.md
@@ -8,6 +8,8 @@
 - D4, D100, D Infinity, Esau Jr, and Dice Bag are now removed from all pools when playing as Random Baby.
 - Lachryphagy is now defined as an anti-synergy with Ipecac. (Thanks Gamonymous)
 - Fixed the bug where the description for long babies would not show up properly in the Rebirth Item Tracker.
+- Changed the following babies:
+  - Hooligan Baby (#514) - The effect no longer applies to bosses.
 - Fixed bugs with the following babies:
   - Slicer Baby (#331) - Now has island softlock prevention.