+++ docs/history.md
@@ -11,6 +11,7 @@
 - Changed the following babies:
   - Hooligan Baby (#514) - The effect no longer applies to bosses.
 - Fixed bugs with the following babies:
+  - Nature Baby (#305) - Now properly reveals the pill effects. (Thanks Gamonymous)
   - Slicer Baby (#331) - Now has island softlock prevention.
 
 ### _v1.20.2_ - September 19th, 2023